Rename forkutils -> forkutil (#1976)
