+++ beacon-chain/chaintest/backend/BUILD.bazel
@@ -20,7 +20,7 @@ go_library(
         "//beacon-chain/utils:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
         "//shared/bls:go_default_library",
-        "//shared/forkutils:go_default_library",
+        "//shared/forkutil:go_default_library",
         "//shared/hashutil:go_default_library",
         "//shared/params:go_default_library",
         "//shared/sliceutil:go_default_library",
