+++ beacon-chain/core/state/BUILD.bazel
@@ -38,7 +38,7 @@ go_test(
         "//beacon-chain/core/helpers:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
         "//shared/bls:go_default_library",
-        "//shared/forkutils:go_default_library",
+        "//shared/forkutil:go_default_library",
         "//shared/hashutil:go_default_library",
         "//shared/params:go_default_library",
         "@com_github_prometheus_client_golang//prometheus/testutil:go_default_library",
