+++ shared/forkutil/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = ["signature.go"],
-    importpath = "github.com/prysmaticlabs/prysm/shared/forkutils",
+    importpath = "github.com/prysmaticlabs/prysm/shared/forkutil",
     visibility = ["//visibility:public"],
     deps = ["//proto/beacon/p2p/v1:go_default_library"],
 )
