+++ shared/forkutil/signature_test.go
@@ -1,4 +1,4 @@
-package forkutils
+package forkutil
 
 import (
 	"math"
